Update on "[WIP] Bucket Batch DataPipe (rename after discussion)"
```python

        dp = IterableWrapper(range(20))
        dp = dp.bucket_batch(batch_size=2, number_of_buckets=5, bucket_function=lambda x: x % 5)
        actual = list(dp)
        expected = [[0, 5], [1, 6], [2, 7], [3, 8], [4, 9], [10, 15], [11, 16], [12, 17], [13, 18], [14, 19]]
        self.assertEqual(expected, actual)

```




[ghstack-poisoned]
